refactor(config): simplify google availability
