+++ packages/backend/src/common/constants/env.constants.test.ts
@@ -1,4 +1,6 @@
 import {
+  ENV,
+  getApiBaseURL,
   isGoogleConfigured,
   parseBackendEnv,
 } from "@backend/common/constants/env.constants";
@@ -18,6 +20,26 @@ const validEnv = {
 };
 
 describe("env.constants", () => {
+  describe("getApiBaseURL", () => {
+    const originalBaseUrl = ENV.BASEURL;
+
+    afterEach(() => {
+      ENV.BASEURL = originalBaseUrl;
+    });
+
+    it("throws a clear error when ENV.BASEURL is blank", () => {
+      ENV.BASEURL = "   ";
+
+      expect(() => getApiBaseURL()).toThrow("ENV.BASEURL is not set");
+    });
+
+    it("returns the original ENV.BASEURL value when set", () => {
+      ENV.BASEURL = " https://api.example.com/api ";
+
+      expect(getApiBaseURL()).toBe(" https://api.example.com/api ");
+    });
+  });
+
   it("parses password-only backend env without Google configuration", () => {
     const env = parseBackendEnv(validEnv);
