+++ packages/backend/src/common/constants/env.constants.ts
@@ -37,6 +37,7 @@ const EnvSchema = z
   .superRefine((env, context) => {
     const hasGoogleClientId = Boolean(env.GOOGLE_CLIENT_ID);
     const hasGoogleClientSecret = Boolean(env.GOOGLE_CLIENT_SECRET);
+    const isGoogleConfigComplete = hasGoogleClientId && hasGoogleClientSecret;
 
     if (hasGoogleClientId !== hasGoogleClientSecret) {
       context.addIssue({
@@ -50,7 +51,7 @@ const EnvSchema = z
     }
 
     if (
-      hasGoogleClientId &&
+      isGoogleConfigComplete &&
       env.BASEURL.startsWith("https://") &&
       !env.TOKEN_GCAL_NOTIFICATION
     ) {
@@ -108,3 +109,11 @@ try {
 
 export const ENV = parsedEnv;
 export const IS_GOOGLE_CONFIGURED = isGoogleConfigured(ENV);
+
+export function getApiBaseURL(): string {
+  if (!ENV.BASEURL?.trim()) {
+    throw new Error("ENV.BASEURL is not set");
+  }
+
+  return ENV.BASEURL;
+}
