+++ packages/backend/src/common/services/gcal/gcal.service.ts
@@ -13,9 +13,9 @@ import {
 } from "@core/types/sync.types";
 import { IDSchemaV4 } from "@core/types/type.utils";
 import { GCAL_PRIMARY } from "@backend/common/constants/backend.constants";
+import { getApiBaseURL } from "@backend/common/constants/env.constants";
 import { error } from "@backend/common/errors/handlers/error.handler";
 import { GcalError } from "@backend/common/errors/integration/gcal/gcal.errors";
-import { getApiBaseURL } from "@backend/common/util/api-base-url.util";
 import { encodeChannelToken } from "@backend/sync/util/watch.util";
 
 class GCalService {
