+++ packages/web/src/common/hooks/useGoogleCmdItems.ts
@@ -1,0 +1,20 @@
+import { type JsonStructureItem } from "react-cmdk";
+import { useConnectGoogle } from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle";
+
+export const useGoogleCmdItems = (): JsonStructureItem[] => {
+  const { commandAction, isAvailable } = useConnectGoogle();
+
+  if (!isAvailable) {
+    return [];
+  }
+
+  return [
+    {
+      id: "connect-google-calendar",
+      children: commandAction.label,
+      icon: commandAction.icon,
+      disabled: commandAction.isDisabled,
+      onClick: commandAction.onSelect,
+    },
+  ];
+};
