+++ packages/web/src/views/Day/components/DayCmdPalette.tsx
@@ -2,10 +2,10 @@ import { useState } from "react";
 import _CommandPalette, { filterItems, getItemIndex } from "react-cmdk";
 import "react-cmdk/dist/cmdk.css";
 import dayjs from "@core/util/date/dayjs";
-import { useConnectGoogle } from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle";
 import { moreCommandPaletteItems } from "@web/common/constants/more.cmd.constants";
 import { VIEW_SHORTCUTS } from "@web/common/constants/shortcuts.constants";
 import { useAuthCmdItems } from "@web/common/hooks/useAuthCmdItems";
+import { useGoogleCmdItems } from "@web/common/hooks/useGoogleCmdItems";
 import { pressKey } from "@web/common/utils/dom/event-emitter.util";
 import {
   openEventFormCreateEvent,
@@ -31,19 +31,8 @@ export const DayCmdPalette = ({ onGoToToday }: DayCmdPaletteProps) => {
   const [page] = useState<"root">("root");
   const [search, setSearch] = useState("");
   const today = dayjs();
-  const { commandAction, isAvailable: isGoogleAvailable } = useConnectGoogle();
   const authCmdItems = useAuthCmdItems();
-  const googleCmdItems = isGoogleAvailable
-    ? [
-        {
-          id: "connect-google-calendar",
-          children: commandAction.label,
-          icon: commandAction.icon,
-          disabled: commandAction.isDisabled,
-          onClick: commandAction.onSelect,
-        },
-      ]
-    : [];
+  const googleCmdItems = useGoogleCmdItems();
 
   const filteredItems = filterItems(
     [
